requirements-trading.txt
flask==3.1.0
flask-cors==5.0.0
anthropic==0.40.0
openai==1.58.1
alpaca-py==0.43.0
python-dotenv==1.0.1
yfinance==0.2.50
pandas>=2.0.0
numpy>=1.24.0

